Fixed dot unstuffing
Read from this in the RFC:

When a line of mail text is received by the SMTP server, it checks
the line.  If the line is composed of a single period, it is
treated as the end of mail indicator.  If the first character is a
period and there are other characters on the line, the first
character is deleted.