Fixes for all HTML bannering cases
This fixes HTML bannering for:

quoted-printable with _no_ </body> or </html>
quoted-printable with </body> or </html>
base64 with _no_ </body> or </html>
base64 with </body> or </html>

I have working tests for all of these cases.  The only anomaly is with
quoted-printable HTML bannering cases.  When we re-encode and assemble
the new buffer, we get an extra newline at the end.  I believe this
is an artifact of quoted-printable encoding.  It does not happen in the
base64 cases.