mail_from.is_resolvable: moved re_bogus_ip into config
was hard coded in the plugin. This allows for local site modification.

This is useful for entities (like me) that have firewalls with all their public IPs. The VMs and services all run on private or loopback IPs, as routed by the firewall. Being behind the firewall, Haraka sees the loopback IPs and rejects them.