note: added geoip, uribl, dnsbl, helo, access, mf.resolv
These plugins are ones that generally did not save connection notes.

Geoip, added a 'human' element to the existing note, with the same summary that is logged.

Helo, added a 'reject' option to the config file, so that the tests can be enabled and I can still get email from Amazon (which fails the is_dynamic test).

DNSBL, added the 'multi' search option (search=all), so that I can log and inspect the results of all the enabled DNSBLs.