results: one bug fix, one new feature, 14 new tests
Bug fix: I tracked down and fixed the cause of NaN in some of my incrementers. It happened when the incrementer was initializing the results store. Tests are added which validate behavior when initialized by add, incr, and push.

New feature: has

Tests a string value or array members for the presence of an exact string match or a pattern match.