dkim_sign: extract signing domain from From header
with fallback to legacy use of env MAIL FROM domain. (In most cases, the header From domain is the same as the envelope MAIL FROM domain. Now that address-rfc2822 is a dependency, attempt to get header From domain as that's what DMARC alignment tests check.)

moved config parsing into register with cb