Don't capture multi-part body text
It doesn't seem particularly useful to capture the multi-part
preamble in Body's text blob, since it's not really a text part.
Also capturing it with any filters applied to the Body was causing
the preamble to be duplicated up the stack in Transaction, due to
Body returning the preamble twice: once as it came in the message,
and later after filters were applied to the captured text.