Update transaction.js
I had a bug when i discarded the data from the rcpt hook, it was crashing as the line was never called and therefore the write_ce.end never existed. checking the line again before calling the callback solved the problem. Tho i'm not sure if there is a better way to do it but this is my current fix.