AVG: result store and add defer options
* use path.join to assemble tmpfile (Windows friendly)
* move config loading into register()
* save results in result_store
* added two config file options
    * [defer]error=false
    * [defer]timeout=false

The motivation for the [defer] options is being able to test deploy this plugin in a safe manner. When there's an AVG engine/connection/network error, I don't want it to block all mail. The default behavior is still to defer on error/timeout, unless the config file options are set.