smtp_forward: revert part of #962 w/overlap #950
two changes:

1. Don't save txn results (redundant with #950)
2. Don't append txn ID to queue message

Question: Is there any reason for the txn ID to be appended? Log
messages are already prefixed with the txn ID and txn results already have
the txn ID available, so at that point, it's entirely redundant data.

Near as I can tell, the only reason to append it there is so that it's
returned to the sending MTA in queue_*_respond. If that's the case, I'd rather
append it there.

Thoughts?

* connection: append uuid *after* saving to results store