Haraka now takes a list of backend forwarding hosts to connect to
The original config for Haraka took a single host and port in the
smtp_forward.ini config to connect to. If that host was a CNAME with a
bunch of possible hosts it could resolve to, Haraka would have no way of
knowing if say just one of the hosts was down.

Now we can have the pool listed explicitly in the config like this:

    forwarding_host_pool=1.1.1.1:1111, 2.2.2.2:2222, 3.3.3.3.:3333....

If that's not in the config, then smtp_client will continue to use the
old host and port settings.

When a host in the pool goes down, we remove it from the pool of available
hosts and we set a timer to check it periodically to see if it looks like
it's come back online.

The first attempted delivery to a failed backend will result in the mail
being rejected back to the originating MTA, who is responsible for
resending it to Haraka.