Call next only after message_stream ended
With the way the plugin was written before, it was possible that a later
plugin would also call connection.transaction.message_stream.pipe()
before dkim_verify's pipe had ended, leading to an error.  This moves
the next() call out of the DKIM callback and into the end handler on the
stream to ensure that'll never happen.

Patch from Matt Sergeant.