Change logerror to logdebug in list detection code
Remove a connection.logerror() from (currently rudimentary) email
list detection code in the data.headers plugin.  As currently coded,
it would record an error to the log file for any email with a "Sender"
header that didn't start with "owner-".

However, rather than deleting entirely, it's been changed to a
logdebug() to help catch any other cases where email list detection
relies on the addition or modification of the Sender header.