Haraka now takes a list of backend hosts to forward to (#1333)
* handle arrays in .ini files

msimerson commented in a pull request https://github.com/haraka/Haraka/pull/1333#issuecomment-184313920

    One thing that might be worth changing is the way the host pool is
    configured. A method that already has widespread use is to specify an array
    for the host value instead of a string. This would necessitate extending our
    ini parser to recognize the array syntax.

The syntax looks like this. There is no "official" .ini syntax for arrays, but
this seems to be common.

    hosts[] = first_host
    hosts[] = second_host
    hosts[] = third_host

This squashed merge from the feature branch also includes

commit e778a5420b4703c2679a291974c334e2daa021bc
Author: Kevin M. Goess <kgoess@craigslist.org>
Date:   Thu Feb 18 09:28:25 2016 -0800

    whitespace fix for eslint

commit 669b0d94ac39dd31612493505516ca38eebe63f0
Author: Kevin M. Goess <kgoess@craigslist.org>
Date:   Thu Feb 18 09:21:50 2016 -0800

    fix test for .ini arrays

    hand-applying msimerson's 86e69a316cd

* Haraka now takes a list of backend forwarding hosts to connect to

The original config for Haraka took a single host and port in the
smtp_forward.ini config to connect to. If that host was a CNAME with a
bunch of possible hosts it could resolve to, Haraka would have no way of
knowing if say just one of the hosts was down.

Now we can have the pool listed explicitly in the config like this:

    forwarding_host_pool=1.1.1.1:1111, 2.2.2.2:2222, 3.3.3.3.:3333....

If that's not in the config, then smtp_client will continue to use the
old host and port settings.

When a host in the pool goes down, we remove it from the pool of available
hosts and we set a timer to check it periodically to see if it looks like
it's come back online.

The first attempted delivery to a failed backend will result in the mail
being rejected back to the originating MTA, who is responsible for
resending it to Haraka.

* s/if/of/ in comment

* default port to 25

msimerson: Since Haraka is a MTA, having a default port 25 (when
unspecified) makes the config syntax a little more friendly.

* lint: 'config' is already declared in upper scope

That's what I get for trying to avoid file-scoped globals.

* cruft from my bad merge

These get_filetype_reader tests were moved to the haraka-config
node_module package. The configfile.js at the root here doesn't
have a get_filetype_reader method any more.