Support + as a wildcard in the aliases plugin
It is common to use '+' as a subaddress separator. With a '+'
based scheme, the aliases plugin now splits out the user part
from the details part (see rfc3598 for terminology) and aliases
the user part as defined in the rules.

Prior to this change, only '-' was supported as a subaddress separator.