outbound: Add results property to transaction
transaction.results is used widely by many plugins to store results.
The code in connection.js creates this property in init_transaction
but similar logic is missing in outbound code.

In the future, results property should move to the transaction constructor.
But this is complicated by the fact that ResultStore requres a connection
object and doing this properly requires a (risky) refactor.

(The specific crash here was that the dkim plugin tries to sign a bounce
mail and crashes when doing a txn.results.add)