delay_deny: add included mode
Currently, this plugin delay denies everything unless listed in the
config file. This can be a little risky when experimenting this plugin
when a lot of other plugins are already in use.

In such cases, it's easier to use an included mode (a whitelist mode)
where one can add plugins to bypass one by one.