Do not create a top level secureContext
This causes problems because the secureContext ends up being used by
Haraka in server and client mode. Currently, the server mode loads
TLS keys as part of the context in (plugins/tls_socket). The client
mode does not load keys (in smtp_client.js). This discrepancy causes
tls connections to fail when there is an incoming mail immediately
followed by an outgoing mail.