Resolve base64 data truncation (#2180)
Under circumstances where base64 data was folded on to lines
that had lengths that were not evenly divisible by 4, creation
of the constituent data buffers would cause between 2-6 bits
to be truncated from each line.

This patch changes decode_base64 so that it will accumulate
up to 3 characters, so that it always emits buffers that
are aligned on byte boundaries.