Implement SIGTERM graceful shutdown if pid is 1 (#2547)
Related: https://blog.ghaiklor.com/avoid-running-nodejs-as-pid-1-under-docker-images-when-running-them-on-mesosphere-kubernetes-or-b7bd505657f9

Like in the blog described node has some signal issues if it runs as pid 1 (usually inside a container).
Currently on docker stop docker sends SIGTERM to node which will be ignored because not handled somewhere.

Haraka on docker or even on kubernetes has to be killed unless you use workarounds like `tini`. I would prefer that haraka initiate a graceful shutdown on docker stop.