haraka.service: Type=simple (#2612)
* haraka.service: Type=simple

`systemctl start hakara.service` was hanging, but the service was started ok. This is because the process wasn't forking and does not appear to have an option to do so. 

https://superuser.com/a/1274913/117296

> a) If the service starts and keeps running, and the prompt does not return until you press Control-C or stop the service in some other way: then Type = simple is the right choice.
> b) If the prompt returns but the service keeps running in the background (i.e. the service daemonizes itself on its own), then Type = forking is the right choice.

* update haraka.service systemd file

Add comment explaining that `daemonize` should be `false`
change recommended path for the unit file - see https://unix.stackexchange.com/a/208352