Fix bannering with nested mimes
This fixes the bug where nested mime structures don't work with
bannering.

This also updates the transaction and mailbody code to work on buffers
instead of always converting to strings, which means we can preserve
exact bytes when we don't need to grok the text coming down the wire.
This part is necessary because the fix involves more manipulation of the
data as it comes in, and converting back and forth to strings to do so
can be lossy.

The change in attachment_stream comes from battle-testing this code and
noticing that workers can stop responding and loop infinitely dequeueing
and re-enqueueing buffered data under certain circumstances.