fix bannering on nested mime parts (#2737)
* Add failing test on bannering with nested mimes

Found in the wild, messages with nested mime structures have a bug where
bannering doesn't work at all.

* Fix bannering with nested mimes

This fixes the bug where nested mime structures don't work with
bannering.

This also updates the transaction and mailbody code to work on buffers
instead of always converting to strings, which means we can preserve
exact bytes when we don't need to grok the text coming down the wire.
This part is necessary because the fix involves more manipulation of the
data as it comes in, and converting back and forth to strings to do so
can be lossy.

The change in attachment_stream comes from battle-testing this code and
noticing that workers can stop responding and loop infinitely dequeueing
and re-enqueueing buffered data under certain circumstances.

* Fix tests to work with new bannering code

Now, turning on body parsing will always normalize newlines.  I've
updated the original czech text test so that it *doesn't* parse the
body, to ensure the bytes come through verbatim.  I've added a second
similar test that *does* parse the body, but allows for normalized
newlines.

* Update Changes.md

* Fix mailbody tests with new buffer interface