Allow web interface to be bound to unix socket
To better support embedded use of Haraka, this change adds support
for specifying listener addresses as /path/to/some.sock[:mode], where
`mode` is an optional octal permissions spec.  Removal of an existing
socket is handled automatically.

This change also consolidates listening-endpoint address parsing and
formatting to a single class, removing duplicated code and simplifying
the setup of servers.