Get local_ip from getsockname() instead of server properties
Fixes #2914

In the case of binding to a specific existing address, client.localAddress
should be the same as the bind address and this commit doesn't change that.

In the case of binding to a non-existing address, no client will exist until
the address does exist, and this commit doesn't change that.

Finally, in the case of binding to INADDR_ANY, the bind address is useless
information (that is: I can't think of any use case where this fact would be
useful during handling an incoming connection), and this commit changes the
value from :: or 0.0.0.0 to a value that is actually useful: the IP address on
which the incoming connection was actually accepted.