outbound: set acquireTimeoutMillis to prevent constant reconnect to unreachable servers
The generic-pool module is built on the assumption that acquire
always succeeds. It is implemented as a busy loop of calling
create() non-stop and there is no way to make acquire() return
an error.

As a mitigation error, we make acquire() fail after 10 seconds.
Note that it will still busy loop for 10 seconds. We have to fix
upstream module or replace generic-pool to really fix the problem.

see #3100