Update helo.checks.js
I found more issues. None of the hooks were getting registered.  There was a for...in loop which loops over the checks array. To get the correct values from the array, it should be a for...of loop. 

Fixing this issue gets the hooks registered, but now each hook get called twice. So now the first time a hook gets called, I check for a previously added result.