Fix crash when outbound connection fails
In node 22 atleast, network code has started returning AggregateError.
This means that when socket connection fails, err.message is an empty string.
The on('error') handler as a result invoked callback('', null) . Since error
is falsy, code assumes socket is valid and crashes.

Code is changed to always return a error string.

Fixes #3455